
\name{expressionSet2TilingSet}
\alias{expressionSet2TilingSet}
\title{Convert an ExpressionSet to a tilingSet object}
\description{
This function converts an ExpressionSet to a tilingSet object from the rMAT package.
}
\usage{
expressionSet2TilingSet(eSet)
}
\arguments{
\item{eSet}{an ExpressionSet}
}
\author{ Benedikt Zacher \email{zacher@lmb.uni-muenchen.de}}
\keyword{manip}
