
\name{intersection}
\alias{intersection}
\title{Get the intersection of two named vectors}
\description{
This function intersects the names of two named vectors and stores the intersection in a matrix.
}
\usage{
intersection(means, expression)
}
\arguments{
\item{means}{a list of named integer vectors.}
\item{expression}{a named integer vector}
}
\value{
a matrix containing the values of features, that are contained in both vectors.
}
\author{ Benedikt Zacher \email{zacher@lmb.uni-muenchen.de}}
\keyword{internal}
