
\name{list2matrix}
\alias{list2matrix}
\title{Convert profile list to matrix}
\description{
This function converts the list of profiles generated by the getProfiles function to a matrix, if all entries have the same length.
}
\usage{
list2matrix(profiles)
}
\arguments{
\item{profiles}{a list, generated by the getProfiles}
}
\value{
a list with with a matrix at the entry profile.
}
\author{ Benedikt Zacher \email{zacher@lmb.uni-muenchen.de}}
\keyword{manip}
