
\name{rankPercentile.normalize}
\alias{rankPercentile.normalize}
\title{Rank precentile Normalization}
\description{
This function performs rank percentile normalization of the data. To use this normalization, the function normalize.Probes should be called with method="rankpercentile".
}
\usage{
rankPercentile.normalize(matrix)
}
\arguments{
\item{matrix}{a matrix, containing the data}
}
\author{ Benedikt Zacher \email{zacher@lmb.uni-muenchen.de}}
\keyword{internal}
