
\name{RGlist2ExpresionSet}
\alias{RGlist2ExpresionSet}
\title{Convert an RGlist to an ExpressionSet}
\description{
This function converts an RGlist to an ExpressionSet.
}
\usage{
RGlist2ExpresionSet(RGlist)
}
\arguments{
\item{RGlist}{an object of class RGlist}
}
\author{ Benedikt Zacher \email{zacher@lmb.uni-muenchen.de}}
\keyword{internal}
