
\name{whichIn}
\alias{whichIn}
\title{Map positions to intervals}
\description{
This function provides efficient mapping of positions to intervals. It is used to calculate the ChIP-profiles of annotated features.
}
\usage{
whichIn(pos,bounds)
}
\arguments{
\item{pos}{a vector, containing the positions}
\item{bounds}{a vector, containing the boundaries}
}
\author{ Benedikt Zacher \email{zacher@lmb.uni-muenchen.de}}
\keyword{internal}
