
\name{windowxy}
\alias{windowxy}
\title{Get mfcol or mfrow for a number of plots to one device}
\description{
This function calculates the parameters for mfcol or mfrow, for a given number of plots.
}
\usage{
windowxy(windows=1)
}
\arguments{
\item{windows}{number of plots}
}
\author{ Benedikt Zacher \email{zacher@lmb.uni-muenchen.de}}
\keyword{internal}
