
\name{writeGFF}
\alias{writeGFF}
\title{write ChIP-chip data to a gff file}
\description{
This function writes the all columns of the assayData to a gff file.
}
\usage{
writeGFF(expressionSet, probeAnno, file)
}
\arguments{
\item{expressionSet}{an ExpressionSet object}
\item{probeAnno}{a probeAnno object}
\item{file}{path to write to}
}
\author{ Benedikt Zacher \email{zacher@lmb.uni-muenchen.de}}
\keyword{IO}
